Add javadocs and remove useless classes
