+++ src/main/java/duke/command/ShowListCommand.java
@@ -4,12 +4,26 @@ import duke.main.Storage;
 import duke.main.TaskList;
 import duke.main.Ui;
 
+/**
+ * A Command which shows the entire TaskList object.
+ */
 public class ShowListCommand extends Command {
+    /**
+     * Prints the entire TaskList object when executed.
+     * @param taskList TaskList object of the Duke object for the Command object to use.
+     * @param ui Ui object of the Duke object for the Command object to use.
+     * @param storage Storage object of the Duke object for the Command object to use.
+     */
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         ui.showList(taskList.getTaskList());
     }
 
+    /**
+     * Whether the MakeToDoCommand should end the Duke object.
+     * It should not.
+     * @return False.
+     */
     @Override
     public boolean isExit() {
         return false;
