+++ src/main/java/duke/exception/DukeException.java
@@ -1,5 +1,8 @@
 package duke.exception;
 
+/**
+ * Parent exception for all duke related exceptions.
+ */
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
