--- src/main/java/duke/exception/InvalidIndexException.java
@@ -1,9 +1,0 @@
-package duke.exception;
-
-import duke.exception.DukeException;
-
-public class InvalidIndexException extends DukeException {
-    public InvalidIndexException(String description) {
-        super(description);
-    }
-}
