--- src/main/java/duke/exception/MissingArgumentException.java
@@ -1,9 +1,0 @@
-package duke.exception;
-
-import duke.exception.DukeException;
-
-public class MissingArgumentException extends DukeException {
-    public MissingArgumentException(String description) {
-        super(description);
-    }
-}
