+++ src/main/java/duke/exception/MissingDescriptionException.java
@@ -1,7 +1,8 @@
 package duke.exception;
 
-import duke.exception.DukeException;
-
+/**
+ * DukeException when a description is missing.
+ */
 public class MissingDescriptionException extends DukeException {
     public MissingDescriptionException(String description) {
         super(description);
