+++ src/main/java/duke/exception/MissingIndexException.java
@@ -1,7 +1,8 @@
 package duke.exception;
 
-import duke.exception.DukeException;
-
+/**
+ * DukeException when an index is missing.
+ */
 public class MissingIndexException extends DukeException {
     public MissingIndexException(String description) {
         super(description);
