+++ src/main/java/duke/exception/MissingTimeException.java
@@ -1,7 +1,8 @@
 package duke.exception;
 
-import duke.exception.DukeException;
-
+/**
+ * DukeException when the time is missing.
+ */
 public class MissingTimeException extends DukeException {
     public MissingTimeException(String description) {
         super(description);
