+++ src/main/java/duke/main/Parser.java
@@ -11,6 +11,14 @@ import java.time.LocalDate;
 
 public class Parser {
 
+    /**
+     * Parses the full command and generates a corresponding Command object which will be executed.
+     * @param fullCommand The full command to be parsed.
+     * @return A Command object which does what the command specifies when executed.
+     * @throws MissingIndexException If the index is needed but missing.
+     * @throws MissingDescriptionException If the description is needed but missing.
+     * @throws MissingTimeException If the time is needed but missing.
+     */
     public static Command parse(String fullCommand)
             throws MissingIndexException, MissingDescriptionException, MissingTimeException {
         String[] splitCommand = fullCommand.split(" ", 2);
@@ -36,6 +44,13 @@ public class Parser {
         }
     }
 
+    /**
+     * Helper method for making the MarkTaskCommand.
+     * @param isDone Whether to mark or unmark the task.
+     * @param command Split command.
+     * @return MarkTaskCommand object that marks or unmarks a task when executed.
+     * @throws MissingIndexException If the index is missing.
+     */
     private static MarkTaskCommand makeMarkTaskCommand(boolean isDone, String[] command) throws MissingIndexException {
         if (command.length < 2) {
             throw new MissingIndexException("Missing index");
@@ -44,6 +59,12 @@ public class Parser {
         return new MarkTaskCommand(isDone, index - 1);
     }
 
+    /**
+     * Helper method for making the DeleteTaskCommand.
+     * @param command Split command.
+     * @return DeleteTaskCommand object that deletes the task at index.
+     * @throws MissingIndexException If the index is missing.
+     */
     private static DeleteTaskCommand makeDeleteTaskCommand(String[] command) throws MissingIndexException {
         if (command.length < 2) {
             throw new MissingIndexException("Missing index");
@@ -52,13 +73,26 @@ public class Parser {
         return new DeleteTaskCommand(index - 1);
     }
 
-    private static MakeToDoTaskCommand makeMakeToDoTaskCommand(String[] command) throws MissingDescriptionException {
+    /**
+     * Helper method for making the MakeToDoCommand.
+     * @param command Split command.
+     * @return MakeToDoTaskCommand object that makes a ToDo object and adds it to a TaskList object when executed.
+     * @throws MissingDescriptionException If the description is missing.
+     */
+    private static MakeToDoCommand makeMakeToDoTaskCommand(String[] command) throws MissingDescriptionException {
         if (command.length < 2) {
             throw new MissingDescriptionException("Missing description");
         }
-        return new MakeToDoTaskCommand(command[1]);
+        return new MakeToDoCommand(command[1]);
     }
 
+    /**
+     * Helper method for making the MakeDeadlineCommand.
+     * @param command Split command.
+     * @return MakeDeadlineCommand object that makes a Deadline object and adds it to a TaskList object when executed.
+     * @throws MissingDescriptionException If the description is missing.
+     * @throws MissingTimeException If the time is missing.
+     */
     private static MakeDeadlineCommand makeMakeDeadlineCommand(String[] command)
             throws MissingDescriptionException, MissingTimeException {
         if (command.length < 2) {
@@ -71,6 +105,13 @@ public class Parser {
         return new MakeDeadlineCommand(splitCommand[0], LocalDate.parse(splitCommand[1]));
     }
 
+    /**
+     * Helper method for making the MakeEventCommand.
+     * @param command Split command.
+     * @return MakeEventObject that makes an Event object and adds it to a TaskList object when executed.
+     * @throws MissingDescriptionException If the description is missing.
+     * @throws MissingTimeException If the time is missing.
+     */
     private static MakeEventCommand makeMakeEventCommand(String[] command)
             throws MissingDescriptionException, MissingTimeException {
         if (command.length < 2) {
